requirements.txt
fastapi
uvicorn[standard]
pydantic
pdf2image
pillow
pytesseract
python-multipart
